requirements.txt
beanie 
datasets 
fastapi 
passlib[bcrypt]
fastapi_mail 
huggingface_hub 
langchain 
langchain_community 
langchain_text_splitters 
motor 
openai 
pandas 
pydantic 
groq
pydantic_settings 
python-dotenv 
redis[hiredis]
tqdm 
uvicorn 
PyJWT
Jinja2
python-multipart
weaviate-client
websockets
minio
langchain-huggingface
pymupdf
tiktoken
transformers
torch
docx2txt
prometheus-fastapi-instrumentator
